+++ lib/src/meteor/subscribed_collection.dart
@@ -3,7 +3,6 @@ import 'package:enhanced_ddp/enhanced_ddp.dart';
 /// Provides useful methods to read data from a collection on the frontend.
 ///
 /// [SubscribedCollection] supports only read functionality useful in case of getting only the data subscribed by user and not any other data.
-/// To access other methods use `Meteor.getCustomDatabase(dbUrl)` and use the methods of the `Db` class.
 class SubscribedCollection {
   /// The internal collection instance.
   final Collection _collection;
@@ -31,7 +30,6 @@ class SubscribedCollection {
   void removeUpdateListeners() {
     _collection.removeUpdateListeners();
   }
-  
 
   /// Returns specific objects from a subscribed collection using a set of [selectors].
   Map<String, Map<String, dynamic>> find(Map<String, dynamic> selectors) {